Fix delete bug with indices
Fixed by preventing the wrappers referencing the same list of elements
by subset classes. The wrappers referenced the same thing. This problem
was made worse as I did not having classes added to the class list on
what I actually wanted to select for anyway, so the list of elements I
wanted would never have been selected for anyway.

So they are the same thing. They reference the list of wrappers. When
the wrappers are updated with new attributes, since that is what is in
the list as items, then they get the updated index but the non-wrappers
still have the old indices originally set as data-ids. The wrappers are
then reset with the new indices.