Track "selected" and "to edit" list separately
Now the selected list strictly tracks the list currently displayed, not
the list edited. This separation helps us check if they are the same
later. If they are the same on edit of a list then change the items.
Otherwise if it is a list currently being displayed there is no point,
no updates to the display need to be done.