On add show entire list of todos without selective addition
This is to prepare for accepting a sorted list of todos, therefore
selective addition is too tedious to do based on how the sorted todo
items move. The function for display will accept the sorted list and
only display the sorted list and not worry about adding selectively
anymore, just depending on the filter or sort released.