Fix all todos except those in default list alone not initially displayed
Default list was always detected but rest of todos from other lists were
missing as they were not stitched up yet. The other lists in the lists
manager did in fact not exist as they had not been added yet. So to fix
the bug I just had to relegate the show of default list all todos to
after all lists with their todos are added to the list manager, not
within the loop while they are yet to be added. In debugging I observed
the default list todos displayed differently at the initial app load and
after when all todos would show as they were now ready and stored in the
system.